// frontend/lib/generated/OrganizationReconStartArgs.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AssetIntelHydrateConfig } from "./AssetIntelHydrateConfig";

export type OrganizationReconStartArgs = { organizationId: string, providerIds: Array<string>, config: AssetIntelHydrateConfig, allowExternal: boolean, allowActive: boolean, };